- fixed Mac OS X paths to binaries - bug report: texturemap not yet supported, instead color surface vertices - patch created with isosurface had single color stored as a scalar in FaceVertexCData (index wrt colormap). In u3d_pre_patch, lines 123-133 test whether FaceVertexCData is either 1 column (indexed) or 3 (already RGB). If indexed, it is expanded to RGB by getting the indexed colors from the colormap using function scaled_ind2rgb in the same file.
However, no test was performed on the number of rows of FaceVertexCData, because these
normally where either equal to the number of vertices (so interpolation of colors from vertex to vertex, shading interp),
or to the number of faces (shading flat).

When ploting a single color surface MATLAB sets FaceVertexCData to a 3x1 matrix.
Using isosurface though results in a scalar color.

Fix applied to u3d_pre_patch as follows.
If a single row FaceVertexCData is encountered, before converting from indexed to
RGB, it is assumed that it corresponds to a single color.
So it does not matter whether CData is the same number of rows with vertices or faces - same result.
The FaceVertexCData gets expanded to the number of rows of the faces matrix.