Add support for hggroups
The code assumed that the parent of a graphics handle is an axes
object. This is not true for more complex graphics object hierarchies
(e.g. involving groups of handles by hggroups). Thus traverse the
graphics tree upwards in case the parent is not an axes until the axes
is finally found.